+++ src/components/InputPrimary/index.test.tsx
@@ -1,0 +1,20 @@
+/**
+ * @prettier
+ */
+
+import * as React from 'react';
+import { noop } from 'lodash';
+
+import render from '~/helpers/test/render';
+
+import InputPrimary from '.';
+
+describe('InputPrimary', () => {
+  it('matches the snapshot', () => {
+    const { container } = render(
+      <InputPrimary onChange={noop} value="Value" />,
+    );
+
+    expect(container).toMatchSnapshot();
+  });
+});