+++ src/components/index.ts
@@ -13,22 +13,28 @@ export * from './Alert';
 export * from './AppBarNavigation';
 export * from './Dialog';
 export * from './EmptyState';
-export * from './FormInput';
-export * from './FormSelect';
+export * from './FormInputGroups';
 export * from './FormSwitch';
-export * from './InputLabel';
 export * from './Loading';
 export * from './SearchBar';
 export * from './ThemeProvider';
 
 // WIP COMPONENTS --------------------------------------------------------------
 
-export * from './FormInputGroups';
+export * from './FormInput';
+export * from './FormSelect';
+export * from './InputLabel';
 export * from './RenderMetaTags';
 
 // DEPRECATED COMPONENTS -------------------------------------------------------
 
 export * from './DataTable';
+export * from './FormInputPrimary';
+export * from './FormSelectPrimary';
+export * from './InputLabelPrimary';
+export * from './InputPrimary';
+export * from './SelectPrimary';
+export * from './SelectSecondary';
 export * from './TableCellBody';
 export * from './TableCellHead';
 export * from './TableContainer';