+++ src/constants/typography.ts
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) ACT, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @prettier
+ */
+
+export default {
+  WEIGHT_BOLD: 700,
+  WEIGHT_LIGHT: 300,
+  WEIGHT_MEDIUM: 600,
+  WEIGHT_REGULAR: 500,
+};