+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Chronological history of changes to the Design Language System.
 
-## [v5.0.0] - June 04, 2021
+## [v5.0.0] - June 07, 2021
 
 * With this release, the DLS will no longer export components that you can
 import directly from Material UI, such as `<Button />` and `<Link />`.